fix(i18n): properly set translation domain during layout init
Most of the `i18n` boilerplate was there, but the app still needed to
know what domain to look for. This must happen after initialization of
the `GtkBuilder` and before loading the glade file.

The domain is simply the `mo` file packaged with the app and usually
located under:

```
/usr/share/locale/$LANG/LC_MESSAGES/$APP.mo
```

Once again thanks to @maximbaz for pointing it out.

Closes #92