Extend desktop file to support screenshot actions
It's super convenient to be able to take a screenshot and pass it to
Swappy in one step.

Actions in the .desktop file feel like a natural place to offer
these options.

Using rofi as a launcher, ".desktop" file actions are findable
if the -drun-show-actions is passed.

With these actions, you can snapshot any of the following and have
the result opened in swappy:

 - current window
 - current output
 - select a window
 - select an area

The downside is that these depend on `grimshot`, which is part of
the `sway` project in the `contrib/` directory there.

Some of the grimshot logic in turn depends on `swaymsg`.

Maybe these desktop actions are better hosted in another project
since this one currently doesn't depend on Sway, but I thought
I would share the idea here to start a conversation.

Another option would be to create a `contrib/grimshot-swappy.desktop`
file. Sway users could opt to install it.

Thanks!