feat(config): fix segfault at exit
In case you don't have custom_color defined in your config file (because
you have old config file, for example), a default value is used. The
default is specified as a string literal and its pointer is assigned to
config->custom_color. The config_free() function calls g_free() on this
pointer uncoditionally but you can't do that on a pointer to string
literal so this causes a segfault.

To avoid that, we can g_strdup() the string literal, just like it is
done for example with config->text_font.