feat(ui): add transparency feature
* feat(ui): add transparency feature

It is sometimes better if the drawing is not fully covering the picture
but is semi-transparent. This patch adds a simple button to enable
transparency. When this button is enabled, everything will be drawn with
50% alpha channel.

* feat(ui): make transparency configurable

This commit adds UI elements and config option for configuring the
transparency value. The default value is 50% but it can be changed in
the 5% - 95% range in 10% steps.

* docs(readme): update README with transparency feature

The previous commit added this info to the man page but README was not
updated. Lets fix this.